+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 # Changelog
 
 ## 0.1.0-alpha.7 [Unreleased]
+### Added
+  - Add new element class `\Fiv\Form\Element\Hidden`
+   
 ### Changed
   - Drop support of the php 5.6 Minimum version is php 7.0
 
 ### Removed
   - #40 Remove old methods: Fiv\Form\Element\Checkbox::unCheck, Fiv\Form\Element\Checkbox::check, Fiv\Form\Element\Checkbox::setValue, Fiv\Form\Element\Checkbox::getValue
   - #45 Remove method `Form::getData`
+
+### Deprecated
+  - Deprecate method `\Fiv\Form\Form::hidden()` See `\Fiv\Form\Element\Hidden`
   
 ## 0.1.0-alpha.6 [2016-10-17]
 ### Added